Add Jaeger resource support to distributed application
This commit introduces a new Jaeger resource to the distributed application builder. It includes the addition of extension methods for configuring the Jaeger resource, along with container image tags and a resource representation. The `GrafanaResourceBuilderExtensions` class has been updated with XML documentation for clarity. New classes for `JaegerBuilderExtensions`, `JaegerContainerImageTags`, and `JaegerResource` have been created to facilitate the integration and management of Jaeger within the application.